+++ Assets/PatchKit Patcher/Scripts/AppData/Local/IUnarchiver.cs
@@ -7,5 +7,11 @@ namespace PatchKit.Unity.Patcher.AppData.Local
         event UnarchiveProgressChangedHandler UnarchiveProgressChanged;
 
         void Unarchive(CancellationToken cancellationToken);
+
+        // set to true to continue unpacking on error. Check HasErrors later to see if there are any
+        bool ContinueOnError { set; }
+
+        // After Unarchive() if set to true, there were unpacking errors.
+        bool HasErrors { get; }
     }
 }