+++ Assets/PatchKit Patcher/Scripts/AppUpdater/AppUpdaterContentStrategy.cs
@@ -17,6 +17,8 @@ namespace PatchKit.Unity.Patcher.AppUpdater
 
         private bool _updateHasBeenCalled;
 
+        public bool RepairOnError { get; set; }
+
         public AppUpdaterContentStrategy(AppUpdaterContext context, UpdaterStatus status)
         {
             Checks.ArgumentNotNull(context, "context");
@@ -25,6 +27,9 @@ namespace PatchKit.Unity.Patcher.AppUpdater
 
             _context = context;
             _status = status;
+
+            // defaults
+            RepairOnError = true;
         }
 
         public StrategyType GetStrategyType()
@@ -102,6 +107,19 @@ namespace PatchKit.Unity.Patcher.AppUpdater
             }
 
             installContent.Execute(cancellationToken);
+
+            if (installContent.NeedRepair && RepairOnError)
+            {
+                DebugLogger.Log("Content installed with errors, requesting repair");
+
+                var appRepairer = new AppRepairer(_context, _status);
+                appRepairer.CheckHashes = true;
+
+                if (!appRepairer.Perform(cancellationToken))
+                {
+                    throw new CannotRepairDiskFilesException("Failed to validate/repair disk files");
+                }
+            }
 
             _context.App.DownloadDirectory.Clear();
         }