+++ Assets/PatchKit Patcher/Scripts/AppUpdater/CannotRepairDiskFilesException.cs
@@ -1,0 +1,9 @@
+namespace PatchKit.Unity.Patcher.AppUpdater
+{
+    public class CannotRepairDiskFilesException : System.Exception
+    {
+        public CannotRepairDiskFilesException(string message) : base(message)
+        {
+        }
+    }
+}